+++ MiKo.Analyzer/Rules/Naming/MiKo_1067_CodeFixProvider.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections.Generic;
+using System.Composition;
+using System.Linq;
+
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CodeFixes;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace MiKoSolutions.Analyzers.Rules.Naming
+{
+    [ExportCodeFixProvider(LanguageNames.CSharp, Name = nameof(MiKo_1067_CodeFixProvider)), Shared]
+    public sealed class MiKo_1067_CodeFixProvider : NamingCodeFixProvider
+    {
+        public override string FixableDiagnosticId => MiKo_1067_PerformMethodsAnalyzer.Id;
+
+        protected override string Title => "Remove 'Perform' from name";
+
+        protected override string GetNewName(ISymbol symbol) => MiKo_1067_PerformMethodsAnalyzer.FindBetterName((IMethodSymbol)symbol);
+
+        protected override SyntaxNode GetSyntax(IReadOnlyCollection<SyntaxNode> syntaxNodes) => syntaxNodes.OfType<MethodDeclarationSyntax>().FirstOrDefault();
+    }
+}
